Image cropping works
I finally got rid of the shitty Cloudinary way, so now the images aren't rotated to all hell. The only issue is that a canvas is used, so when cropping out of a high quality image the quality goes to shit. So once I fix that the cropping will be 100% perfected.