Image blob no longer uploads
The problem where the image blob would be uploaded because the user saved too quickly is gone, and is replaced with a loading icon so they will wait to hit the save button. If I so desire, I can even try disabling the save button until the upload is completed in the future.