Preparing for public
I'm never going to have the time or motivation to finish this project as long as I'm the only one with access to the code base. Hopefully opening this up will let others teach me a thing or two about what I've been doing wrong and will inspire me to make it better.

I honestly hate the new banner icon, so I'm going to come back and change that real soon.