Migration from expo
I *think* the migration was successful? Everything seems to be in order and the functionality is working on the Android emulator on my PC, which has been useful because it shows issues with scaling and floating problems the footer on smaller screens. I need to see if I can still test it on my actual mobile and the like, but we have functionality, so I'm deeming this commit-worthy.